digihack_behavior/requirements.txt
flask
ultralytics
opencv-python-headless
pillow
werkzeug